FIX Filesystem API

+++ src/ignfs.cpp
@@ -8,9 +8,23 @@ ignfs::ignfs(QObject *parent) :
 
 }
 
-bool ignfs::fileRemove(const QString &path){
-    QFile file(path);
-    return file.remove();
+bool ignfs::rm(const QString &path){
+    QVariantMap info = this->info(path).toMap();
+    QString absolutePath = info["absoluteFilePath"].toString();
+    if(info["isFile"].toBool()){
+        QFile file(absolutePath);
+        return file.remove();
+    }
+    if(info["isDir"].toBool()){
+        QDir dir;
+        return dir.rmdir(absolutePath);
+    }
+    return false;
+}
+
+bool ignfs::mkdir(const QString &path){
+    QDir dir;
+    return dir.mkdir(path);
 }
 
 bool ignfs::fileWrite(const QString &path, const QString &data){
@@ -59,128 +73,11 @@ QString ignfs::homePath(){
     return home;
 }
 
-bool ignfs::dir(const QString &opt, const QString &path){
-    QDir dir;
-    if(opt == "create"){
-        return dir.mkdir(path);
-    }
-    else if(opt == "remove"){
-        return dir.rmdir(path);
-    }
-    else{
-        return false;
-    }
-}
-
-bool ignfs::isExist(const QString &path)
-{
-    return QFile::exists(path);
-}
-
-bool ignfs::isDirectory(const QString &path)
-{
-    return QFileInfo(path).isDir();
-}
-
-bool ignfs::isFile(const QString &path)
-{
-    return QFileInfo(path).isFile();
-}
-
-bool ignfs::isAbsolute(const QString &path)
-{
-   return QFileInfo(path).isAbsolute();
-}
-
-bool ignfs::isExecutable(const QString &path)
-{
-   return QFileInfo(path).isExecutable();
-}
-
-bool ignfs::isSymlink(const QString &path)
-{
-   return QFileInfo(path).isSymLink();
-}
-
-bool ignfs::isReadable(const QString &path)
-{
-   return QFileInfo(path).isReadable();
-}
-
-bool ignfs::isWritable(const QString &path)
-{
-   return QFileInfo(path).isWritable();
-}
-
 bool ignfs::copy(const QString &src, const QString &des){
     if(QFile::exists(des)){
         QFile::remove(des);
     }
     return QFile::copy(src,des);
-}
-
-QString ignfs::openFileDialog(){
-    QFileDialog *fd = new QFileDialog;
-    int result = fd->exec();
-    if (result)
-    {
-        QString directory = fd->selectedFiles()[0];
-        return directory;
-    }
-    else {
-        return "";
-    }
-}
-
-QString ignfs::openDirDialog(){
-    QFileDialog *fd = new QFileDialog;
-#ifdef Linux
-    QTreeView *tree = fd->findChild <QTreeView*>();
-    tree->setRootIsDecorated(true);
-    tree->setItemsExpandable(true);
-#endif
-    fd->setFileMode(QFileDialog::Directory);
-    fd->setOption(QFileDialog::ShowDirsOnly);
-    fd->setViewMode(QFileDialog::Detail);
-    int result = fd->exec();
-    QString directory;
-    if (result)
-    {
-        directory = fd->selectedFiles()[0];
-        return directory;
-    }
-    else {
-        return "";
-    }
-}
-
-QString ignfs::saveFileDialog(){
-    QFileDialog *fd = new QFileDialog;
-    QString directory = fd->getSaveFileName();
-    return directory;
-}
-
-QString ignfs::saveFileDialog(const QVariant &config){
-    QVariantMap conf = json->jsonParser(config).toVariantMap();
-    QString title = "Save File",
-            path = this->homePath(),
-            ext = "";
-    if(conf["title"].toString() != ""){
-        title = conf["title"].toString();
-    }
-
-    if(conf["path"].toString() != ""){
-        path = conf["path"].toString();
-    }
-
-    if(conf["type"].toString() != ""){
-        ext = conf["type"].toString();
-    }
-
-    QFileDialog *fd = new QFileDialog;
-    QWidget *widget = new QWidget();
-    QString directory = fd->getSaveFileName(widget, title,path,ext);
-    return directory;
 }
 
 QStringList ignfs::list(const QString &path){
